refactor: organized S3 upload code
