+++ src/config/index.js
@@ -1,6 +1,6 @@
 import { Kafka } from 'kafkajs';
-import AWS from 'aws-sdk';
 import api from '../api.js';
+import { uploadLogsToS3 } from '../utils/uploadLogs.js';
 
 const client = new Kafka({
   clientId: 'logger',
@@ -18,35 +18,8 @@ await admin.createTopics({
 });
 await admin.disconnect();
 
-const s3 = new AWS.S3({
-  accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-  region: process.env.AWS_REGION,
-  endpoint: `https://s3.${process.env.AWS_REGION}.amazonaws.com`,
-  s3ForcePathStyle: true,
-});
-
-async function uploadLogsToS3(requestId, logs) {
-  console.log('Uploading logs to S3...');
-  const fileContent = JSON.stringify(logs, null, 2);
-  const fileName = `logs/${requestId}-${Date.now()}.json`;
-
-  const params = {
-    Bucket: process.env.S3_BUCKET_NAME,
-    Key: fileName,
-    Body: fileContent,
-    ContentType: 'application/json',
-  };
-
-  try {
-    const result = await s3.upload(params).promise();
-    console.log(`Logs uploaded to S3: ${result.Location}`);
-  } catch (error) {
-    console.error(`Error uploading logs to S3: ${error.message}`);
-  }
-}
-
 const logsByRequestId = {};
+const pendingTimers = {};
 
 await consumer.connect()
 await consumer.subscribe({ topics: ['microservice-logs', 'gateway-logs'] })
@@ -55,20 +28,27 @@ await consumer.run({
     console.log(message.value.toString())
 
     const log = JSON.parse(message.value.toString());
-    const { message: logMessage, params: { request_id: requestId } } = log;
 
+    if (log.params.request_id) {
+      const { params: { request_id: requestId } } = log;
 
-    if (requestId) {
       if (!logsByRequestId[requestId]) {
         logsByRequestId[requestId] = [];
       }
 
       logsByRequestId[requestId].push(log);
 
-      if (logMessage === 'Connection closed') {
+      // If there is a pending timer for this request, clear it
+      if (pendingTimers[requestId]) {
+        clearTimeout(pendingTimers[requestId]);
+      }
+
+      // Configurate a timer to upload logs to S3 in 5 seconds
+      pendingTimers[requestId] = setTimeout(async () => {
         await uploadLogsToS3(requestId, logsByRequestId[requestId]);
         delete logsByRequestId[requestId];
-      }
+        delete pendingTimers[requestId];
+      }, 5000);
     }
   },
 })
